requirements-dev.txt
# Development requirements (additional packages for development only)
-r requirements.txt

# Testing
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==6.0.0
factory-boy==3.3.1

# Code quality
flake8==7.1.1
black==24.10.0
isort==5.13.2

# Documentation
sphinx==8.1.3
